fix(bflp): cast operand to long in audit trim subtraction (sonar S2184)
Promote nodes.size() to long before the subtraction so the arithmetic
happens in 64-bit, satisfying SonarQube S2184 (implicit narrow cast of
subtraction result).

+++ src/main/java/org/jahia/modules/bruteforceloginprotection/core/AuditLogger.java
@@ -189,7 +189,7 @@ public class AuditLogger {
         if (nodes.size() <= max) {
             return;
         }
-        long toRemove = nodes.size() - max;
+        long toRemove = (long) nodes.size() - max;
         nodes.sort((a, b) -> {
             try {
                 long ta = a.hasProperty(PROP_AUDIT_TIMESTAMP) ? a.getProperty(PROP_AUDIT_TIMESTAMP).getLong() : 0L;
